Add week 6 review material

+++ src/spring-2195/week05/multiple-files/shapes.c
@@ -1,5 +1,5 @@
 #include "shapes.h"
 
-int area(int height, int width) {
+float area(int height, int width) {
     return height * width;
 }
