+++ src/spring-2195/week05/multiple-files/shapes.h
@@ -1,6 +1,6 @@
 #ifndef SHAPES_H
 #define SHAPES_H
 
-int area(int height, int width);
+float area(int height, int width);
 
 #endif
